+++ 2D/Ising-cftp.cpp
@@ -17,8 +17,8 @@ class IsingCFTP : public Bitmap<int> {
 public:
     explicit IsingCFTP(const Hub &H)
         : Bitmap<int>(H.title, {H['n'], H['n']}), b(H['b']), d(0), s(H['s']), status(H.title, {size_t(w()), size_t(h())}) {
-        for (int i = 0; i < w(); ++i)
-            for (int j = 0; j < h(); ++j) put({i, j}, 1);
+        for (size_t i = 0; i < w(); ++i)
+            for (size_t j = 0; j < h(); ++j) put({i, j}, 1);
         snap();
         b *= log(1 + sqrt(double(2)));
         for (int i = 0; i <= 4; ++i) p.push_back(exp(b * i) / (exp(b * i) + exp(b * (4 - i))));
@@ -54,8 +54,8 @@ public:
         int n = w() * h();
         while (n > 0) {
             cerr << n << endl;
-            for (int i = d; i < w() - 2 * d; ++i)
-                for (int j = d; j < h() - 2 * d; ++j) put({i, j}, 1);
+            for (size_t i = d; i < w() - 2 * d; ++i)
+                for (size_t j = d; j < h() - 2 * d; ++j) put({i, j}, 1);
             for (auto t = states.size(); t-- > 0;) {
                 prng.state(states[t]);
                 for (unsigned i = 0; i < (1u << t); ++i) up();
@@ -72,29 +72,29 @@ public:
 
     void bc_0() {
         d = 1;
-        for (int i = 0; i < w(); ++i) {
+        for (size_t i = 0; i < w(); ++i) {
             put({i, 0}, 0);
-            put({i, h() - 1}, 0);
+            put({i, size_t(h()) - 1}, 0);
         }
-        for (int j = 0; j < h(); ++j) {
+        for (size_t j = 0; j < h(); ++j) {
             put({0, j}, 0);
-            put({w() - 1, j}, 0);
+            put({size_t(w()) - 1, j}, 0);
         }
     }
 
     void bc_dobrushin() {
         d = 1;
-        for (int i = 0; i < w(); ++i) {
+        for (size_t i = 0; i < w(); ++i) {
             put({i, 0}, 0);
-            put({i, h() - 1}, 2);
+            put({i, size_t(h()) - 1}, 2);
         }
-        for (int j = 0; j < h() / 2; ++j) {
+        for (size_t j = 0; j < h() / 2; ++j) {
             put({0, j}, 0);
-            put({w() - 1, j}, 0);
+            put({size_t(w()) - 1, j}, 0);
         }
-        for (int j = h() / 2; j < h(); ++j) {
+        for (size_t j = h() / 2; j < h(); ++j) {
             put({0, j}, 2);
-            put({w() - 1, j}, 2);
+            put({size_t(w()) - 1, j}, 2);
         }
     }
