+++ 2D/Snake2.cpp
@@ -20,19 +20,19 @@ class Snake : public Bitmap<site> {
 public:
     Snake(const Hub &H, size_t n, double l, bool hex)
         : Bitmap(H.title, {4 * n + 1, 2 * n + 1}), lambda(l), path(1, {2 * int(n), 0}), hex(hex) {
-        for (int x = 0; x < w(); ++x) put({x, 0}, VERTEX);
+        for (size_t x = 0; x < size_t(w()); ++x) put({x, 0u}, VERTEX);
         if (double a = H['a']; a != 0) triangle(a);
         if (!H['v']) show();
     }
 
     void triangle(double a) {
         double ta = tan(a * M_PI / 180);
-        for (int y = 0; y < h(); ++y) {
+        for (size_t y = 0; y < h(); ++y) {
             int  yy = y + (y % 2);
             auto s  = int(2 + yy * ta);
             s += (s % 2);
-            for (int x = 0; x < w(); ++x)
-                if (abs(x - w() / 2) >= s) put({x, y}, VERTEX);
+            for (size_t x = 0; x < w(); ++x)
+                if (abs(int(x) - w() / 2) >= s) put({x, y}, VERTEX);
         }
     }
